refactor: Create types for modules
